--- src/main/java/com/finalproject/recruit/jisoo/dto/NoticeRecruitsRes.java
@@ -1,29 +1,0 @@
-package com.finalproject.recruit.jisoo.dto;
-
-import com.finalproject.recruit.entity.Recruit;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@Builder
-public class NoticeRecruitsRes {
-        private Long recruitId;
-        private String recruitTitle;
-        private String recruitProcedure;
-        private LocalDateTime createdTime;
-
-        public static NoticeRecruitsRes fromRecruit(Recruit recruit){
-            return NoticeRecruitsRes.builder()
-                    .recruitId(recruit.getRecruitId())
-                    .recruitTitle(recruit.getRecruitTitle())
-                    .recruitProcedure(recruit.getRecruitProcedure())
-                    .createdTime(recruit.getCreatedTime())
-                    .build();
-
-        }
-
-}
