--- src/main/java/com/finalproject/recruit/jisoo/parameter/Keywords.java
@@ -1,20 +1,0 @@
-package com.finalproject.recruit.jisoo.parameter;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum Keywords {
-    keyword1,
-    keyword2,
-    keyword3,
-    keyword4,
-    keyword5,
-    keyword6,
-    keyword7,
-    keyword8,
-    keyword9,
-    keyword10
-
-}
