--- src/main/java/com/finalproject/recruit/jisoo/parameter/MilitaryEnum.java
@@ -1,7 +1,0 @@
-package com.finalproject.recruit.jisoo.parameter;
-
-public enum MilitaryEnum {
-    면제,
-    군필,
-    미필
-}
