--- src/main/java/com/finalproject/recruit/jisoo/repository/apply/ActivitiesRepository.java
@@ -1,8 +1,0 @@
-package com.finalproject.recruit.jisoo.repository.apply;
-
-import com.finalproject.recruit.entity.apply.Activities;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ActivitiesRepository extends JpaRepository<Activities, Long> {
-
-}
