--- src/main/java/com/finalproject/recruit/jisoo/repository/apply/AwardsRepository.java
@@ -1,8 +1,0 @@
-package com.finalproject.recruit.jisoo.repository.apply;
-
-import com.finalproject.recruit.entity.apply.Awards;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AwardsRepository extends JpaRepository<Awards, Long> {
-
-}
