--- src/main/java/com/finalproject/recruit/jisoo/repository/apply/CareerRepository.java
@@ -1,8 +1,0 @@
-package com.finalproject.recruit.jisoo.repository.apply;
-
-import com.finalproject.recruit.entity.apply.Career;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CareerRepository extends JpaRepository<Career, Long> {
-
-}
