--- src/main/java/com/finalproject/recruit/jisoo/repository/apply/CertificateRepository.java
@@ -1,8 +1,0 @@
-package com.finalproject.recruit.jisoo.repository.apply;
-
-import com.finalproject.recruit.entity.apply.Certificate;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CertificateRepository extends JpaRepository<Certificate, Long> {
-
-}
