--- src/main/java/com/finalproject/recruit/jisoo/repository/apply/EducationRepository.java
@@ -1,8 +1,0 @@
-package com.finalproject.recruit.jisoo.repository.apply;
-
-import com.finalproject.recruit.entity.apply.Education;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface EducationRepository extends JpaRepository<Education, Long> {
-
-}
