--- src/main/java/com/finalproject/recruit/jisoo/repository/apply/LanguageRepository.java
@@ -1,8 +1,0 @@
-package com.finalproject.recruit.jisoo.repository.apply;
-
-import com.finalproject.recruit.entity.apply.Language;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LanguageRepository extends JpaRepository<Language, Long> {
-
-}
