--- src/main/java/com/finalproject/recruit/jisoo/repository/apply/MilitaryRepository.java
@@ -1,8 +1,0 @@
-package com.finalproject.recruit.jisoo.repository.apply;
-
-import com.finalproject.recruit.entity.apply.Military;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MilitaryRepository extends JpaRepository<Military, Long> {
-
-}
