+++ src/main/java/com/finalproject/recruit/parameter/MilitaryEnum.java
@@ -1,0 +1,7 @@
+package com.finalproject.recruit.parameter;
+
+public enum MilitaryEnum {
+    면제,
+    군필,
+    미필
+}
