+++ src/main/java/com/finalproject/recruit/repository/apply/ActivitiesRepository.java
@@ -1,0 +1,8 @@
+package com.finalproject.recruit.repository.apply;
+
+import com.finalproject.recruit.entity.apply.Activities;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ActivitiesRepository extends JpaRepository<Activities, Long> {
+
+}
