+++ src/main/java/com/finalproject/recruit/repository/apply/AwardsRepository.java
@@ -1,0 +1,8 @@
+package com.finalproject.recruit.repository.apply;
+
+import com.finalproject.recruit.entity.apply.Awards;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AwardsRepository extends JpaRepository<Awards, Long> {
+
+}
