+++ src/main/java/com/finalproject/recruit/repository/apply/CareerRepository.java
@@ -1,0 +1,8 @@
+package com.finalproject.recruit.repository.apply;
+
+import com.finalproject.recruit.entity.apply.Career;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CareerRepository extends JpaRepository<Career, Long> {
+
+}
