+++ src/main/java/com/finalproject/recruit/repository/apply/CertificateRepository.java
@@ -1,0 +1,8 @@
+package com.finalproject.recruit.repository.apply;
+
+import com.finalproject.recruit.entity.apply.Certificate;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CertificateRepository extends JpaRepository<Certificate, Long> {
+
+}
