+++ src/main/java/com/finalproject/recruit/repository/apply/EducationRepository.java
@@ -1,0 +1,8 @@
+package com.finalproject.recruit.repository.apply;
+
+import com.finalproject.recruit.entity.apply.Education;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface EducationRepository extends JpaRepository<Education, Long> {
+
+}
