+++ src/main/java/com/finalproject/recruit/repository/apply/LanguageRepository.java
@@ -1,0 +1,8 @@
+package com.finalproject.recruit.repository.apply;
+
+import com.finalproject.recruit.entity.apply.Language;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LanguageRepository extends JpaRepository<Language, Long> {
+
+}
