+++ src/main/java/com/finalproject/recruit/repository/apply/MilitaryRepository.java
@@ -1,0 +1,8 @@
+package com.finalproject.recruit.repository.apply;
+
+import com.finalproject.recruit.entity.apply.Military;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MilitaryRepository extends JpaRepository<Military, Long> {
+
+}
